fix(tests): move computeToggledFavorites to a pure-utility module
The previous test file imported computeToggledFavorites from the hook
file, which dragged the whole React + i18n side-effect chain (including
syncHtmlLang touching `document` at top level) into the test runner —
fine in JSDOM, but the project's vitest config runs in node, so CI
crashed with `ReferenceError: document is not defined`.

Extract the pure helper to src/utils/recipeFavorites.ts so its test can
import without booting the React context.
