+++ src/utils/recipeFavorites.ts
@@ -1,0 +1,7 @@
+/** Pure: returns the new set after toggling `key`. */
+export function computeToggledFavorites(current: ReadonlySet<string>, key: string): Set<string> {
+  const next = new Set(current);
+  if (next.has(key)) next.delete(key);
+  else next.add(key);
+  return next;
+}
